cmam/requirements.txt
future
numpy
matplotlib
pandas
scipy
nested_dict
dicthash
joblib
